Correctly namespace and autoload DatabaseException
